refactor(test): consolidate duplicate helper functions into testutils_test.go
- Create testutils_test.go with shared helper functions
- Add generic containsInSlice[T] function
- Remove duplicate helper functions from:
  - session_test.go (newTestLogger)
  - runner_test.go (contains, containsHelper)
  - session_manager_test.go (containsInSlice)
  - websocket_test.go (containsStr -> contains)
  - provider_test.go (assertContains)
- Clean up unused imports

Refs #9
