+++ runner_test.go
@@ -305,20 +305,6 @@ func TestEngine_DangerDetectorMethods(t *testing.T) {
 	}
 }
 
-// Helper function
-func contains(s, substr string) bool {
-	return len(s) >= len(substr) && (s == substr || len(s) > 0 && containsHelper(s, substr))
-}
-
-func containsHelper(s, substr string) bool {
-	for i := 0; i <= len(s)-len(substr); i++ {
-		if s[i:i+len(substr)] == substr {
-			return true
-		}
-	}
-	return false
-}
-
 // mockSessionManager for testing
 type mockSessionManager struct {
 	sessions map[string]*Session
